+++ api/src/main/java/com/yavlash/api/exception/InvalidInputException.java
@@ -1,4 +1,4 @@
-package com.yavlash.api.exceptions;
+package com.yavlash.api.exception;
 
 public class InvalidInputException extends RuntimeException{
     public InvalidInputException() {
